requirements.txt
polars>=0.19.0
mysql-connector-python>=8.0.33
nba_api>=1.4.1
python-dotenv>=1.0.0
pyarrow>=12.0.0
